requirements.txt
# Core
openai>=1.0.0
anthropic>=0.18.0
pyyaml>=6.0
requests>=2.28.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
rawpy>=0.18.0
exifread>=3.0.0

# Video Processing
moviepy>=1.0.3

# Utilities
python-magic>=0.4.27
tqdm>=4.65.0
colorlog>=4.8.0
